// src/utils/types/enums/documentEnums.ts
export enum PhysicalExamData {
  Weight = 'Weight',
  WeightMeasurement = 'WeightMeasurement',
  Height = 'Height',

  SystolicPressure = 'SystolicPressure',
  DiastolicPressure = 'DiastolicPressure',
  HeartRate = 'HeartRate',
  RespiratoryRate = 'RespiratoryRate',
  Temperature = 'Temperature',
  Complaints = 'Complaints',

  InspectionComplete = 'InspectionComplete',
  InspectionDeviation = 'InspectionDeviation',
  InspectionIndicator = 'InspectionIndicator',
  InspectionComment = 'InspectionComment',

  SkinComplete = 'SkinComplete',
  SkinDeviation = 'SkinDeviation',
  SkinIndicator = 'SkinIndicator',
  SkinComment = 'SkinComment',

  MucousComplete = 'MucousComplete',
  MucousDeviation = 'MucousDeviation',
  MucousIndicator = 'MucousIndicator',
  MucousComment = 'MucousComment',

  LymphComplete = 'LymphComplete',
  LymphDeviation = 'LymphDeviation',
  LymphIndicator = 'LymphIndicator',
  LymphComment = 'LymphComment',

  EyesEarsNoseComplete = 'EyesEarsNoseComplete',
  EyesEarsNoseDeviation = 'EyesEarsNoseDeviation',
  EyesEarsNoseIndicator = 'EyesEarsNoseIndicator',
  EyesEarsNoseComment = 'EyesEarsNoseComment',

  HeadNeckComplete = 'HeadNeckComplete',
  HeadNeckDeviation = 'HeadNeckDeviation',
  HeadNeckIndicator = 'HeadNeckIndicator',
  HeadNeckComment = 'HeadNeckComment',

  UpperVeinsComplete = 'UpperVeinsComplete',
  UpperVeinsDeviation = 'UpperVeinsDeviation',
  UpperVeinsIndicator = 'UpperVeinsIndicator',
  UpperVeinsComment = 'UpperVeinsComment',

  CardiovascularComplete = 'CardiovascularComplete',
  CardiovascularDeviation = 'CardiovascularDeviation',
  CardiovascularIndicator = 'CardiovascularIndicator',
  CardiovascularComment = 'CardiovascularComment',

  RespiratoryComplete = 'RespiratoryComplete',
  RespiratoryDeviation = 'RespiratoryDeviation',
  RespiratoryIndicator = 'RespiratoryIndicator',
  RespiratoryComment = 'RespiratoryComment',

  MusculoskeletalComplete = 'MusculoskeletalComplete',
  MusculoskeletalDeviation = 'MusculoskeletalDeviation',
  MusculoskeletalIndicator = 'MusculoskeletalIndicator',
  MusculoskeletalComment = 'MusculoskeletalComment',

  DigestiveComplete = 'DigestiveComplete',
  DigestiveDeviation = 'DigestiveDeviation',
  DigestiveIndicator = 'DigestiveIndicator',
  DigestiveComment = 'DigestiveComment',

  UrinaryComplete = 'UrinaryComplete',
  UrinaryDeviation = 'UrinaryDeviation',
  UrinaryIndicator = 'UrinaryIndicator',
  UrinaryComment = 'UrinaryComment',

  NervousComplete = 'NervousComplete',
  NervousDeviation = 'NervousDeviation',
  NervousIndicator = 'NervousIndicator',
  NervousComment = 'NervousComment',

  MentalComplete = 'MentalComplete',
  MentalDeviation = 'MentalDeviation',
  MentalIndicator = 'MentalIndicator',
  MentalComment = 'MentalComment',

  ECGTaken = 'ECGTaken',
  ECGTapeAttached = 'ECGTapeAttached',
  ECGComment = 'ECGComment',

  UrineAnalysis = 'UrineAnalysis',
  UrineAnalysisName = 'UrineAnalysisName',
  UrineAnalysisResult = 'UrineAnalysisResult',
  UrineKotininAnalysis = 'UrineKotininAnalysis',
  UrineKotininAnalysisResult = 'UrineKotininAnalysisResult',
  AlcoholTestPromiles = 'AlcoholTestPromiles',
  PregnancyTest = 'PregnancyTest',
  PregnancyTestResult = 'PregnancyTestResult',
  CovidTest = 'CovidTest',
  CovidTestResult = 'CovidTestResult',

  CovidOldTest = 'CovidOldTest',
  CovidOldTestDate = 'CovidOldTestDate',
  Test10HoursAware = 'Test10HoursAware',

  LabDate = 'LabDate',
  EmptyStomach = 'EmptyStomach',
  BloodSamples = 'BloodSamples',
  BloodCentralLab = 'BloodCentralLab',
  UrineCentralLab = 'UrineCentralLab',
  LabRoute = 'LabRoute',
  LabRouteName = 'LabRouteName',
  RgSurveyRoute = 'RgSurveyRoute',
  XRay = 'XRay',
  XRayCopy = 'XRayCopy',
  Analysis = 'Analysis',
  AnalysisDate = 'AnalysisDate',

  GeneralBloodComplete = 'GeneralBloodComplete',
  GeneralBloodIndicator = 'GeneralBloodIndicator',
  GeneralBloodDeviation = 'GeneralBloodDeviation',
  GeneralBloodComment = 'GeneralBloodComment',

  BioBloodComplete = 'BioBloodComplete',
  BioBloodIndicator = 'BioBloodIndicator',
  BioBloodDeviation = 'BioBloodDeviation',
  BioBloodComment = 'BioBloodComment',

  UrinalysisSedimentComplete = 'UrinalysisSedimentComplete',
  UrinalysisSedimentIndicator = 'UrinalysisSedimentIndicator',
  UrinalysisSedimentDeviation = 'UrinalysisSedimentDeviation',
  UrinalysisSedimentComment = 'UrinalysisSedimentComment',

  HIVComplete = 'HIVComplete',
  HIVIndicator = 'HIVIndicator',
  HIVDeviation = 'HIVDeviation',
  HIVComment = 'HIVComment',
}